servicex_app/pyproject.toml
[tool.poetry]
name = "servicex-app"
version = "1.0.0-RC.3"
homepage = "https://iris-hep.org"
license = "BSD"
maintainers = ["ServiceX Team <bengal1@illinois.edu>"]
authors = []
description = "REST Frontend to ServiceX"
readme = "README.rst"

packages = [{include = "servicex_app"}]

[tool.poetry.dependencies]
python = "~3.10"
markupsafe = "^2.1.1"
jinja2 = "^3.1.4"
flask = "^2.3.3"
werkzeug = "^3.0.6"
flask-wtf = "^1.0.1"
flask-cors = "^4.0.1"
wtforms = "^3.0.1"
email-validator = "^1.3.0"
Celery = "^5.2.0"
tenacity = "^9.0.0"
pika = "^1.3.1"
flask-restful = "^0.3.9"
flask-jwt-extended = "^4.4.4"
passlib = "^1.7.4"
flask-sqlalchemy = "^3.0.2"
kubernetes = "^25.3.0"
globus-sdk = "^3.13.0"
cryptography = "^43.0.1"
bootstrap-flask = "^2.1.0"
blinker = "^1.5"
pre-commit = "^2.20.0"
minio = "^7.1.12"
flask-migrate = "^3.1.0"
psycopg2 = "^2.9.5"
python-logstash = "^0.4.8"
humanize = "^4.4.0"
gunicorn = "^22.0.0"
requests-toolbelt = "^0.10.1"
urllib3 = "1.26.19"
requests = "^2.31"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
flake8 = "^5.0.4"
pytest-mock = "^3.10.0"
pytest = "^7.2.0"
pytest-flask = "^1.2.0"
coverage = "^6.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ["servicex_app"]
